address race condition in visitor supersession
we occasionally see two supersessions occur concurrently for the same
visitor. one of those will succeed and the other will fail. this results
in the loss of the supersession history. this is not really a huge deal,
but the annoying thing is that this results in an inactionable error.

the solution in this changeset is to use savepoints and rescuing to
swallow the unique violations.