Don't record assignment events for feature gates (#69)
* Don't record assignment events for feature gates

Because feature gates are not used for analysis, there's no need to
record anything but overrides for feature gates. This will vastly reduce
the data that needs to change when deciding a feature gate split. It
also will cause reweighting feature gates to affect clients in the field
instantly unless they've been overridden.

Also explain the difference between experiments and gates to admins.

* more clarity on why specific variants are assigned

* preposition

* Let's do this.

* :cop:

* don't use the word population for gates

because we use it in the admin to mean "recorded assignees of experiments."