Introduce Split Registry Endpoint Accepting a Timestamp (#133)
* Initial pass

* add specs

* split_registry_snapshots

After discussion we decided to rename this endpoint and fold the build_timestamp validation logic into SplitRegistry. Besides making this more of a true resource route, it explains what this endpoint is retunring much better, and consolidates code. Now the SplitRegistrySnapshot class accepts a timestamp and returns to you the split registry as of that time. So this change will be backwards compatifble as we just need to pass it `Time.zone.now`

* v2/builds/:timestamp/split_registry

* uses build_timestamp param

* upgrade to v3

* combine endpoint naming approaches

* remove references to timestamp and split registry snapshot

* final pr code review tweaks

Co-authored-by: Alex Burgel <aburgel@gmail.com>