Add Optional Environment Banner (#212)
Currently, if an organization uses Test Track for multiple environments
(e.g. production and stage), it can be a bit difficult to discern at
first glance for which environment a user is managing a given split
test.

This proposed change gives users the option to set a
`DEPLOYMENT_ENV_LABEL` variable in the `.env` or `.env.local` file,
which will subsequently be rendered in a bright red banner, like so:

![Screenshot 2024-03-20 at 5 23
52 PM](https://github.com/Betterment/test_track/assets/67171899/f273fb03-dd86-433d-bd7b-77268bf27ae5)


The banner is intentionally a tad obnoxious to make it glaringly obvious
to the user which environment they are in.

When no `DEPLOYMENT_ENV_LABEL` variable is set, the header renders as
usual:

![Screenshot 2024-03-20 at 5 23
24 PM](https://github.com/Betterment/test_track/assets/67171899/fc01c995-b5d0-40ac-ac97-1421a725c449)


I envision organizations using the existing state for production and the
banner state for a non-production environment (or perhaps vice-versa).