drop 3.1 from the ci matrix
So, I'm shooting first and asking questions later here. Reason being
that with_transactional_lock doesn't have a version for which we can
simultaneously have Rails 7.1 with Ruby 3.1 (arbitrarily; we just
happened to drop <3.2 before adding 7.1 support).

Since this is an app (instead of a gem) and since this is meant to be
pulled in a deployed as a unit anyways I _think_ I can get away with
this but :shrug: