feat: Upgrade to Rails 7.1 (#238)
### Summary

This bumps Rails to 7.1 and addresses deprecation warnings, with an
exception. It also drops Ruby 3.1 testing and bumps rubocop's
`TargetRubyVersion` to 3.2 (all resulting lints were `rubocop -a`'d,
since they were `Style/HashSyntax`).

I'm leaving opting into 7.1 defaults and running `app:update` to a
follow-up.

### Other Information
There's a couple of deprecation warnings left to deal with, namely:

```
DEPRECATION WARNING: `Rails.application.secrets` is deprecated in favor of `Rails.application.credentials` 
and will be removed in Rails 7.2.
```
and

```
DEPRECATION WARNING: Your `secret_key_base` is configured in `Rails.application.secrets`, 
which is deprecated in favor of `Rails.application.credentials` and will be removed in Rails 7.2.
```

These both come from `devise`, which has a `SecretKeyFinder`:
https://github.com/heartcombo/devise/blob/v4.9.4/lib/devise/secret_key_finder.rb.
That'll check if it can find the secret key in the different possible
configs (`secrets` and `credentials`) and call the appropriate method.
Since we're using `secrets`, we'll trigger the deprecation warning.

This app doesn't call `secrets` explicitly itself, and we rely (AFAICT)
on setting the `SECRET_KEY_BASE` env var. Since
`Rails::Application#secret_key_base` ultimately looks for that, I think
we can at least punt on the secrets -> credentials migration.

/domain @Betterment/test_track_core