bundle update aws-sdk-core (#242)
### Summary

Adds support for AWS authentication with EKS Pod Identity. To support
eks pod identity `aws-sdk-core` needs to be updated to at least version
[`3.188.0`](https://github.com/aws/aws-sdk-ruby/blob/version-3/gems/aws-sdk-core/CHANGELOG.md#31880-2023-11-22)
which brings in the following feature.

> Feature - Support AWS_CONTAINER_AUTHORIZATION_TOKEN_FILE in
ECSCredentials and also allow for ECS and EKS link-local http addresses.

In this PR, I ran `bundle update aws-sdk-core --conservative` which
updated the gem and related gems.

> Thanks for contributing to TestTrack!

/domain @Betterment/test_track_core