Add admin page to show current test track assignments
For longer running experiments with a specific audience in my mind (like
preferred beta testers) it's useful to use the bulk assignments
capability to enable those features. However if there are mulitple
instances of "adding users to the beta", one doesn't have visibility
into what assignments may already be present.  This view is intended to
give admins more confidence that the intended assignments are in place.