Add admin page to show current test track assignments (#249)
### Summary

For longer running experiments with a specific audience in my mind (like
preferred beta testers) it's useful to use the bulk assignments
capability to enable those features. However if there are mulitple
instances of "adding users to the beta", one doesn't have visibility
into what assignments may already be present. This view is intended to
give admins more confidence that the intended assignments are in place.

### Other Information

I've arbitrarily selected a limit of 1000 assignments to avoid
performance issues for splits that have massive bulk assignments of huge
populations. This is entirely too many records to skim IMO, but I didn't
want to take on pagination as part of this minor quality of life
improvement.

This is checked via a system spec, but creating those assignments
results in at least 5000 inserts and adds about 5 seconds to the test. I
debated testing with passing the limit as an optional query parameter,
but wasn't sure that would give the necessary confidence that TT
wouldn't allow a user to exceed the limit. But maybe that's ok -
allowing `?limit=any` while defaulting to a small number of records
would allow for a speedier test and allow for use cases with larger data
sets.

**Note**: Rendering the page locally is quite quick (median < 200ms).

### Screenshots

![Screenshot 2025-04-10 at 4 21
29 PM](https://github.com/user-attachments/assets/0fa15bd6-8904-498e-ab2d-77beef078627)
![Screenshot 2025-04-10 at 4 10
20 PM](https://github.com/user-attachments/assets/c294f6b8-7c8e-4c02-89db-970643e91e8c)

/domain @Betterment/test_track_core

---------

Co-authored-by: Ivan Greene <ivan@ivan.sh>